Move base class in internal package
